+++ Wrappers/Python/cil/optimisation/functions/MixedL21Norm.py
@@ -62,8 +62,8 @@ def _proximal_step_numpy(arr, tau):
     
     Parameters:
     -----------
-    tmp : DataContainer/ numpy array, best if contiguous memory. 
-    tau: float or DataContainer
+    arr : DataContainer/ numpy array, best if contiguous memory. 
+    tau: float, numpy array or DataContainer
 
     Returns:
     --------